+++ backend/src/main/java/com/woowacourse/moamoa/auth/service/AuthService.java
@@ -1,6 +1,7 @@
 package com.woowacourse.moamoa.auth.service;
 
 import com.woowacourse.moamoa.auth.infrastructure.JwtTokenProvider;
+import com.woowacourse.moamoa.auth.service.oauthclient.OAuthClient;
 import com.woowacourse.moamoa.auth.service.response.GithubProfileResponse;
 import com.woowacourse.moamoa.auth.service.response.TokenResponse;
 import com.woowacourse.moamoa.member.service.MemberService;
