+++ backend/src/main/java/com/woowacourse/moamoa/auth/service/oauthclient/GithubOAuthClient.java
@@ -1,4 +1,4 @@
-package com.woowacourse.moamoa.auth.service;
+package com.woowacourse.moamoa.auth.service.oauthclient;
 import com.woowacourse.moamoa.auth.service.request.AccessTokenRequest;
 import com.woowacourse.moamoa.auth.service.response.GithubProfileResponse;
 import com.woowacourse.moamoa.auth.service.response.OAuthAccessTokenResponse;
@@ -12,7 +12,7 @@ import org.springframework.stereotype.Component;
 import org.springframework.web.client.RestTemplate;
 
 @Component
-public class OAuthClient {
+public class GithubOAuthClient implements OAuthClient {
 
     private static final String ACCESS_TOKEN_URL = "https://github.com/login/oauth/access_token";
     private static final String PROFILE_URL = "https://api.github.com/user";
@@ -21,7 +21,7 @@ public class OAuthClient {
     private final String clientSecret;
     private final RestTemplate restTemplate;
 
-    public OAuthClient(
+    public GithubOAuthClient(
             @Value("${oauth2.github.client-id}") final String clientId,
             @Value("${oauth2.github.client-secret}") final String clientSecret,
             final RestTemplate restTemplate) {
@@ -30,6 +30,7 @@ public class OAuthClient {
         this.restTemplate = restTemplate;
     }
 
+    @Override
     public String getAccessToken(final String code) {
         final AccessTokenRequest accessTokenRequest = new AccessTokenRequest(clientId, clientSecret, code);
 
@@ -37,7 +38,9 @@ public class OAuthClient {
         headers.add(HttpHeaders.ACCEPT, MediaType.APPLICATION_JSON_VALUE);
 
         final HttpEntity<AccessTokenRequest> httpEntity = new HttpEntity<>(accessTokenRequest, headers);
-        final OAuthAccessTokenResponse accessTokenResponse = restTemplate.exchange(ACCESS_TOKEN_URL, HttpMethod.POST, httpEntity, OAuthAccessTokenResponse.class).getBody();
+        final OAuthAccessTokenResponse accessTokenResponse = restTemplate.exchange(
+                ACCESS_TOKEN_URL, HttpMethod.POST, httpEntity, OAuthAccessTokenResponse.class
+        ).getBody();
 
         if (Objects.isNull(accessTokenResponse)) {
             throw new IllegalStateException("Access Token을 가져올 수 없습니다.");
@@ -45,6 +48,7 @@ public class OAuthClient {
         return accessTokenResponse.getAccessToken();
     }
 
+    @Override
     public GithubProfileResponse getProfile(final String accessToken) {
         HttpHeaders headers = new HttpHeaders();
         headers.add("Accept", MediaType.APPLICATION_JSON_VALUE);
