+++ backend/src/main/java/com/woowacourse/moamoa/auth/service/oauthclient/OAuthClient.java
@@ -1,0 +1,9 @@
+package com.woowacourse.moamoa.auth.service.oauthclient;
+
+import com.woowacourse.moamoa.auth.service.response.GithubProfileResponse;
+
+public interface OAuthClient {
+
+    String getAccessToken(final String code);
+    GithubProfileResponse getProfile(final String accessToken);
+}
